+++ applications/vessel-history/src/data/config.ts
@@ -7,6 +7,10 @@ export const WORKSPACE_ENV =
   (process.env.NODE_ENV as WorkspaceEnv) ||
   'production'
 
+export const DEBUG_MODE =
+  ((process.env.REACT_APP_DEBUG_MODE ?? false) as boolean) ||
+  (WORKSPACE_ENV === 'production' ? false : true)
+
 export const FULL_SUFIX = 'full'
 
 export const DEFAULT_WORKSPACE: AppState = {
@@ -51,3 +55,5 @@ export const EVENTS_COLORS: Record<string, string> = {
   fishing: '#fff',
   fishingLabels: '#163f89',
 }
+
+export const DEFAULT_EMPTY_VALUE = ' --- '