+++ applications/vessel-history/src/features/profile/components/Info.tsx
@@ -5,7 +5,9 @@ import { useSelector } from 'react-redux'
 import ImageGallery from 'react-image-gallery'
 import { DateTime } from 'luxon'
 import { Button, IconButton } from '@globalfishingwatch/ui-components'
+import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { VesselWithHistory } from 'types'
+import I18nDate from 'features/i18n/i18nDate'
 import { selectCurrentOfflineVessel } from 'features/vessels/offline-vessels.selectors'
 import { useOfflineVesselsAPI } from 'features/vessels/offline-vessels.hook'
 import { OfflineVessel } from 'types/vessel'
@@ -90,63 +92,63 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
             )}
             <div className={styles.identifiers}>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.name}
                 value={vessel.shipname}
                 valuesHistory={vessel.history.shipname.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.type}
-                value={vessel.type ?? ''}
+                value={vessel.type}
                 valuesHistory={vessel.history.vesselType.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.flag}
                 value={vessel.flag}
                 valuesHistory={vessel.history.flag.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.mmsi}
-                value={vessel.mmsi ?? ''}
+                value={vessel.mmsi}
                 valuesHistory={vessel.history.mmsi.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.imo}
-                value={vessel.imo ?? ''}
+                value={vessel.imo}
                 valuesHistory={vessel.history.imo.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.callsign}
-                value={vessel.callsign ?? ''}
+                value={vessel.callsign}
                 valuesHistory={vessel.history.callsign.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.geartype}
-                value={vessel.geartype ?? ''}
+                value={vessel.geartype}
                 valuesHistory={vessel.history.geartype.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.length}
-                value={vessel.length ?? ''}
+                value={vessel.length}
                 valuesHistory={vessel.history.length.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.grossTonnage}
-                value={vessel.grossTonnage ?? ''}
+                value={vessel.grossTonnage}
                 valuesHistory={vessel.history.grossTonnage.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.depth}
-                value={vessel.depth ?? ''}
+                value={vessel.depth}
                 valuesHistory={vessel.history.depth.byDate}
               ></InfoField>
               <div className={styles.identifierField}>
@@ -159,22 +161,46 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
                 )}
               </div>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.builtYear}
-                value={vessel.builtYear ?? ''}
+                value={vessel.builtYear}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.owner}
-                value={vessel.owner ?? ''}
+                value={vessel.owner}
                 valuesHistory={vessel.history.owner.byDate}
               ></InfoField>
               <InfoField
-                vesselName={vessel.shipname ?? ''}
+                vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.operator}
-                value={vessel.operator ?? ''}
+                value={vessel.operator}
                 valuesHistory={vessel.history.operator.byDate}
               ></InfoField>
+
+              <div className={styles.identifierField}>
+                <label>{t(`vessel.aisTransmission_plural`, 'AIS Transmissions')}</label>
+                <div>
+                  {vessel.firstTransmissionDate || vessel.lastTransmissionDate ? (
+                    <Fragment>
+                      {t('common.from', 'from')}{' '}
+                      {vessel.firstTransmissionDate ? (
+                        <I18nDate date={vessel.firstTransmissionDate} />
+                      ) : (
+                        DEFAULT_EMPTY_VALUE
+                      )}{' '}
+                      {t('common.to', 'to')}{' '}
+                      {vessel.lastTransmissionDate ? (
+                        <I18nDate date={vessel.lastTransmissionDate} />
+                      ) : (
+                        DEFAULT_EMPTY_VALUE
+                      )}
+                    </Fragment>
+                  ) : (
+                    DEFAULT_EMPTY_VALUE
+                  )}
+                </div>
+              </div>
             </div>
           </Fragment>
         )}