+++ applications/vessel-history/src/features/profile/components/InfoFieldHistory.tsx
@@ -1,6 +1,7 @@
 import React, { Fragment, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Modal } from '@globalfishingwatch/ui-components'
+import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { ValueItem } from 'types'
 import I18nDate from 'features/i18n/i18nDate'
 import { VesselFieldLabel } from './InfoField'
@@ -59,11 +60,17 @@ const InfoFieldHistory: React.FC<ListItemProps> = ({
               </div>
               <div className={styles.identifierField}>
                 <label>{t('common.currentTimeRange', 'CURRENT TIME RANGE')}</label>
-                <div>
-                  <span className={styles.rangeLabel}>{t('common.since', 'Since')}: </span>
-                  <span className={styles.rangeValue}>{since && <I18nDate date={since} />}</span>
-                  <span className={styles.rangeValue}>{!since && ' - '}</span>
-                </div>
+                {since && (
+                  <div>
+                    <span className={styles.rangeLabel}>{t('common.since', 'Since')}: </span>
+                    <span className={styles.rangeValue}>{<I18nDate date={since} />}</span>
+                  </div>
+                )}
+                {!since && (
+                  <div>
+                    <span className={styles.rangeValue}>{DEFAULT_EMPTY_VALUE}</span>
+                  </div>
+                )}
               </div>
             </div>
             {previousHistory.map((historyValue: ValueItem, index) => (