+++ applications/vessel-history/src/features/profile/components/activity/ActivityModalContent.tsx
@@ -1,13 +1,10 @@
 import React, { Fragment, useMemo } from 'react'
-import { DateTime } from 'luxon'
-import { useTranslation } from 'react-i18next'
 import { EventTypes } from '@globalfishingwatch/api-types'
-import I18nDate from 'features/i18n/i18nDate'
 import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
 import ActivityModalContentDetailsFishing from './ActivityModalContentDetailsFishing'
 import ActivityModalContentDetailsLoitering from './ActivityModalContentDetailsLoitering'
-import styles from './ActivityModalDetails.module.css'
 import ActivityModalContentDetailsEncounter from './ActivityModalContentDetailsEncounter'
+import styles from './ActivityModalDetails.module.css'
 
 interface ActivityModalContentProps {
   event: RenderedEvent
@@ -15,8 +12,6 @@ interface ActivityModalContentProps {
 
 const ActivityModalContent: React.FC<ActivityModalContentProps> = (props): React.ReactElement => {
   const event = props.event
-  const { t } = useTranslation()
-
   const detailsPerType = useMemo(() => {
     switch (event.type) {
       case EventTypes.Fishing:
@@ -42,52 +37,7 @@ const ActivityModalContent: React.FC<ActivityModalContentProps> = (props): React
 
   return (
     <Fragment>
-      <div className={styles.modalContainer}>
-        <p>
-          <span>{t('event.start', 'Start date')}:</span>{' '}
-          <I18nDate date={event.start} format={DateTime.DATETIME_FULL} />
-        </p>
-        <p>
-          <span>{t('event.end', 'End date')}:</span>{' '}
-          <I18nDate date={event.end} format={DateTime.DATETIME_FULL} />
-        </p>
-        <p>
-          <span>{t('event.avgDuration', 'Avg Duration')}:</span> {event.durationDescription}
-        </p>
-        {detailsPerType}
-        <p>
-          <span>{t('event.startDistanceShore', 'Start distance from shore')}:</span>{' '}
-          {event.distances?.startDistanceFromShoreKm
-            ? t('event.formatDistanceKm', '{{value}} km', {
-              value: event.distances?.startDistanceFromShoreKm?.toFixed(2),
-            })
-            : '-'}
-        </p>
-        <p>
-          <span>{t('event.endDistanceShore', 'End distance from shore')}:</span>{' '}
-          {event.distances?.endDistanceFromShoreKm
-            ? t('event.formatDistanceKm', '{{value}} km', {
-              value: event.distances?.endDistanceFromShoreKm?.toFixed(2),
-            })
-            : '-'}
-        </p>
-        <p>
-          <span>{t('event.startDistancePort', 'Start distance from port')}:</span>{' '}
-          {event.distances?.startDistanceFromPortKm
-            ? t('event.formatDistanceKm', '{{value}} km', {
-              value: event.distances?.startDistanceFromPortKm?.toFixed(2),
-            })
-            : '-'}
-        </p>
-        <p>
-          <span>{t('event.endDistancePort', 'End distance from port')}:</span>{' '}
-          {event.distances?.endDistanceFromPortKm
-            ? t('event.formatDistanceKm', '{{value}} km', {
-              value: event.distances?.endDistanceFromPortKm?.toFixed(2),
-            })
-            : '-'}
-        </p>
-      </div>
+      <div className={styles.modalContainer}>{detailsPerType}</div>
     </Fragment>
   )
 }