+++ applications/vessel-history/src/store.ts
@@ -4,6 +4,7 @@ import offlineVesselsReducer from 'features/vessels/offline-vessels.slice'
 import vesselsReducer from 'features/vessels/vessels.slice'
 import searchReducer from 'features/search/search.slice'
 import settingsReducer from 'features/settings/settings.slice'
+import filtersReducer from 'features/profile/filters/filters.slice'
 import { initializeDataviews } from 'features/dataviews/dataviews.utils'
 import mapReducer from './features/map/map.slice'
 import dataviewsReducer from './features/dataviews/dataviews.slice'
@@ -23,6 +24,7 @@ const rootReducer = combineReducers({
   vessels: vesselsReducer,
   search: searchReducer,
   settings: settingsReducer,
+  filters: filtersReducer,
   location: location,
   map: mapReducer,
   dataviews: dataviewsReducer,